+++ src/makefile.vc
@@ -60,7 +60,7 @@ support = \
 	pj_internal.obj
 
 pipeline = \
-	pj_obs_api.obj PJ_cart.obj PJ_pipeline.obj PJ_horner.obj PJ_helmert.obj \
+	proj_4D_api.obj PJ_cart.obj PJ_pipeline.obj PJ_horner.obj PJ_helmert.obj \
 	PJ_vgridshift.obj PJ_hgridshift.obj PJ_unitconvert.obj PJ_molodensky.obj
 
 geodesic = geodesic.obj
